+++ functions/index.js
@@ -12,7 +12,7 @@ exports.api = require('./api')
 
 // functions for direct calls
 const { submitTalk, unsubmitTalk } = require('./direct/submission')
-const { delivered } = require('./direct/emailWebhooks')
+const { delivered } = require('./email/emailWebhooks')
 
 exports.submitTalk = submitTalk
 exports.unsubmitTalk = unsubmitTalk
